+++ CHANGELOG.md
@@ -6,6 +6,8 @@ _Unreleased_
 
 **[DE]**
 - Die Strafe "Einen Monat Kerker" hat nun eine variable Länge, abhängig von der Schwere der Schuld. Die Auswirkung auf die Gesundheit des verurteilten Spielers ist dann entsprechend höher.
+- Methode "TestamentVollstrecken" hinzugefügt (ausgelagert aus WinForms Client)
+- Kleinere Optimierungen
 
 ## 2.1.1