+++ CHANGELOG.md
@@ -7,6 +7,7 @@ _Unreleased_
 **[DE]**
 - Die Strafe "Einen Monat Kerker" hat nun eine variable Länge, abhängig von der Schwere der Schuld. Die Auswirkung auf die Gesundheit des verurteilten Spielers ist dann entsprechend höher.
 - Methode "TestamentVollstrecken" hinzugefügt (ausgelagert aus WinForms Client)
+- Rohstoffrechte (Handelszertifikate) werden nun bei jeder Amtsverleihung sowie Kauf einer Zollburg oder eines Räuberlagers gewährt und sind nicht mehr abhängig von Talergrenzen sowie nicht mehr beschränbkt auf maximal 5.
 - Kleinere Optimierungen
 
 ## 2.1.1